Netflix/requirements.txt
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
scikit-learn==1.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
streamlit==1.29.0
jupyter==1.0.0
jupyterlab==4.0.9
papermill==2.4.0
mlxtend==0.23.0
apyori==1.1.2
wordcloud==1.9.2
folium==0.15.1
ipywidgets==8.1.1
nbformat==5.9.2
python-dotenv==1.0.0 